pip-requirements.txt
Django==1.5.1
django-extensions==1.1.1
South==0.7.6
MySQL-python==1.2.4
ipython==0.13.2
Fabric==1.6.0
django-cms==2.4.2
cmsplugin-embedded-pages==0.0.4
django-filer==0.9.5
cmsplugin-contact==1.0.0
django-haystack==2.1.0
Whoosh==2.5.1
gunicorn==0.17.4
recaptcha-client==1.0.6
cmsplugin-fbalbum==0.5.6
cmsplugin-gallery==0.5.3
cmsplugin-flickr-slideshow==0.1
django-tinymce==1.5.1
pillow
git+https://github.com/hovel/imagestore.git
hg+https://abarcia@bitbucket.org/abarcia/cmsplugin-youtube
django-registration
